requirements.txt
matplotlib==3.6.2
openpyxl==3.0.10
psycopg2-binary==2.9.3
streamlit == 1.24.0
sqlalchemy==1.4.40
xlsxwriter==3.0.3
seawater==3.3.4